GEODE-9807: Removed delegate class to simplify as suggested by review
