+++ geode-gfsh/src/main/java/org/apache/geode/management/internal/cli/commands/StopGatewaySenderCommand.java
@@ -15,7 +15,16 @@
 
 package org.apache.geode.management.internal.cli.commands;
 
+import static org.apache.geode.logging.internal.executors.LoggingExecutors.newCachedThreadPool;
+
+import java.util.ArrayList;
+import java.util.Iterator;
+import java.util.List;
 import java.util.Set;
+import java.util.concurrent.Callable;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Future;
 import java.util.function.BiFunction;
 import java.util.function.Supplier;
 
@@ -27,25 +36,34 @@ import org.apache.geode.distributed.DistributedMember;
 import org.apache.geode.management.cli.CliMetaData;
 import org.apache.geode.management.cli.ConverterHint;
 import org.apache.geode.management.cli.GfshCommand;
+import org.apache.geode.management.internal.SystemManagementService;
 import org.apache.geode.management.internal.cli.result.model.ResultModel;
+import org.apache.geode.management.internal.cli.result.model.TabularResultModel;
 import org.apache.geode.management.internal.i18n.CliStrings;
 import org.apache.geode.management.internal.security.ResourceOperation;
 import org.apache.geode.security.ResourcePermission;
 
 public class StopGatewaySenderCommand extends GfshCommand {
-  private Supplier<StopGatewaySenderCommandDelegate> commandDelegateSupplier;
+  private final ExecutorService executorService;
+  private SystemManagementService managementService;
   private BiFunction<String[], String[], Set<DistributedMember>> findMembers;
+  private final Supplier<StopGatewaySenderOnMember> stopperOnMemberFactory;
 
   @SuppressWarnings("unused") // invoked by spring shell
   public StopGatewaySenderCommand() {
-    this(null, null);
+    this(newCachedThreadPool("Stop Sender Command Thread ", true),
+        StopGatewaySenderOnMemberWithBeanImpl::new, null, null);
     findMembers = this::findMembers;
-    commandDelegateSupplier = this::getCommandDelegate;
   }
 
-  StopGatewaySenderCommand(Supplier<StopGatewaySenderCommandDelegate> commandDelegateSupplier,
+  StopGatewaySenderCommand(
+      ExecutorService executorService,
+      Supplier<StopGatewaySenderOnMember> stopperOnMemberFactory,
+      SystemManagementService managementService,
       BiFunction<String[], String[], Set<DistributedMember>> findMembers) {
-    this.commandDelegateSupplier = commandDelegateSupplier;
+    this.executorService = executorService;
+    this.stopperOnMemberFactory = stopperOnMemberFactory;
+    this.managementService = managementService;
     this.findMembers = findMembers;
   }
 
@@ -72,17 +90,62 @@ public class StopGatewaySenderCommand extends GfshCommand {
       return ResultModel.createError(CliStrings.NO_MEMBERS_FOUND_MESSAGE);
     }
 
-    return commandDelegateSupplier.get().executeStopGatewaySender(senderId.trim(), getCache(),
-        dsMembers);
+    return executeStopGatewaySender(senderId.trim(), getCache(), dsMembers);
   }
 
-  private StopGatewaySenderCommandDelegate getCommandDelegate() {
-    return new StopGatewaySenderCommandDelegateParallelImpl(getManagementService());
+  public ResultModel executeStopGatewaySender(String id, Cache cache,
+      Set<DistributedMember> dsMembers) {
+    List<DistributedMember> dsMembersList = new ArrayList<>(dsMembers);
+    List<Callable<List<String>>> callables = new ArrayList<>();
+    if (managementService == null) {
+      managementService = getManagementService();
+    }
+    for (final DistributedMember member : dsMembersList) {
+      callables.add(() -> stopperOnMemberFactory.get()
+          .executeStopGatewaySenderOnMember(id,
+              cache, managementService, member));
+    }
+
+    List<Future<List<String>>> futures;
+    try {
+      futures = executorService.invokeAll(callables);
+    } catch (InterruptedException ite) {
+      Thread.currentThread().interrupt();
+      return ResultModel.createError(
+          CliStrings.format(CliStrings.GATEWAY_SENDER_STOP_0_COULD_NOT_BE_INVOKED_DUE_TO_1, id,
+              ite.getMessage()));
+    } finally {
+      executorService.shutdown();
+    }
+
+    return buildResultModelFromMembersResponses(id, dsMembersList, futures);
+  }
+
+  private ResultModel buildResultModelFromMembersResponses(String id,
+      List<DistributedMember> dsMembers, List<Future<List<String>>> futures) {
+    ResultModel resultModel = new ResultModel();
+    TabularResultModel resultData = resultModel.addTable(CliStrings.STOP_GATEWAYSENDER);
+    Iterator<DistributedMember> memberIterator = dsMembers.iterator();
+    for (Future<List<String>> future : futures) {
+      DistributedMember member = memberIterator.next();
+      List<String> memberStatus;
+      try {
+        memberStatus = future.get();
+        resultData.addMemberStatusResultRow(memberStatus.get(0),
+            memberStatus.get(1), memberStatus.get(2));
+      } catch (InterruptedException | ExecutionException ite) {
+        resultData.addMemberStatusResultRow(member.getId(),
+            CliStrings.GATEWAY_ERROR,
+            CliStrings.format(CliStrings.GATEWAY_SENDER_0_COULD_NOT_BE_STOPPED_ON_MEMBER_DUE_TO_1,
+                id, ite.getMessage()));
+      }
+    }
+    return resultModel;
   }
 
   @FunctionalInterface
-  interface StopGatewaySenderCommandDelegate {
-    ResultModel executeStopGatewaySender(String id, Cache cache, Set<DistributedMember> dsMembers);
+  interface StopGatewaySenderOnMember {
+    List<String> executeStopGatewaySenderOnMember(String id, Cache cache,
+        SystemManagementService managementService, DistributedMember member);
   }
-
 }
