+++ geode-gfsh/src/test/java/org/apache/geode/management/internal/cli/commands/StopGatewaySenderCommandTest.java
@@ -15,30 +15,71 @@
 
 package org.apache.geode.management.internal.cli.commands;
 
-import static org.apache.geode.management.internal.cli.commands.StopGatewaySenderCommand.StopGatewaySenderCommandDelegate;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.doReturn;
+import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collection;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Set;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Future;
 import java.util.function.BiFunction;
 import java.util.function.Supplier;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
+import org.junit.Before;
 import org.junit.ClassRule;
 import org.junit.Test;
+import org.mockito.ArgumentCaptor;
 
+import org.apache.geode.cache.Cache;
 import org.apache.geode.distributed.DistributedMember;
+import org.apache.geode.management.internal.SystemManagementService;
 import org.apache.geode.management.internal.cli.result.model.ResultModel;
+import org.apache.geode.management.internal.cli.result.model.TabularResultModel;
+import org.apache.geode.management.internal.i18n.CliStrings;
 import org.apache.geode.test.junit.rules.GfshParserRule;
 
 public class StopGatewaySenderCommandTest {
   @ClassRule
   public static GfshParserRule gfsh = new GfshParserRule();
+
+  private final String senderId = "sender1";
+  private Cache cache;
+  private Set<DistributedMember> members;
+  @SuppressWarnings("unchecked")
+  BiFunction<String[], String[], Set<DistributedMember>> findMembers =
+      mock(BiFunction.class);
+  private ExecutorService executorService;
+  private SystemManagementService managementService;
+  private Supplier<StopGatewaySenderCommand.StopGatewaySenderOnMember> stopperOnMemberFactory;
+
+  @SuppressWarnings("unchecked")
+  @Before
+  public void setUp() {
+    cache = mock(Cache.class);
+    members =
+        Stream.generate(() -> mock(DistributedMember.class)).limit(3)
+            .collect(Collectors.toSet());
+
+    executorService = mock(ExecutorService.class);
+    managementService = mock(SystemManagementService.class);
+    stopperOnMemberFactory = mock(Supplier.class);
+    doReturn(mock(StopGatewaySenderCommand.StopGatewaySenderOnMember.class))
+        .when(stopperOnMemberFactory).get();
+    when(findMembers.apply(any(), any())).thenReturn(members);
+  }
 
   @Test
   public void whenMissingSenderIdCommandReturnsInvalidCommandError() {
@@ -51,10 +92,13 @@ public class StopGatewaySenderCommandTest {
   public void whenNoMembersCommandReturnsNoMembersError() {
     // arrange
     @SuppressWarnings("unchecked")
-    BiFunction<String[], String[], Set<DistributedMember>> findMembers = mock(BiFunction.class);
+    BiFunction<String[], String[], Set<DistributedMember>> findMembers =
+        mock(BiFunction.class);
     Set<DistributedMember> emptySet = new HashSet<>();
     when(findMembers.apply(any(), any())).thenReturn(emptySet);
-    StopGatewaySenderCommand command = new StopGatewaySenderCommand(null, findMembers);
+    StopGatewaySenderCommand command =
+        new StopGatewaySenderCommand(executorService, stopperOnMemberFactory,
+            managementService, findMembers);
 
     // act and assert
     gfsh.executeAndAssertThat(command, "stop gateway-sender --id=sender1")
@@ -63,28 +107,60 @@ public class StopGatewaySenderCommandTest {
 
   @SuppressWarnings("unchecked")
   @Test
-  public void whenParamsPassedOkCommandReturnsResultFromDelegate() {
+  public void stopGatewaySenderStartsOneThreadPerMemberAndBuildsOutputAccordingToFuturesOutput()
+      throws InterruptedException, ExecutionException {
     // arrange
-    Set<DistributedMember> members =
-        Stream.generate(() -> mock(DistributedMember.class)).limit(3)
-            .collect(Collectors.toSet());
-    BiFunction<String[], String[], Set<DistributedMember>> findMembers = mock(BiFunction.class);
-    when(findMembers.apply(any(), any())).thenReturn(members);
-    StopGatewaySenderCommandDelegate commandDelegate = mock(StopGatewaySenderCommandDelegate.class);
-    Supplier<StopGatewaySenderCommandDelegate> commandDelegateSupplier = mock(Supplier.class);
-    when(commandDelegateSupplier.get()).thenReturn(commandDelegate);
-    ResultModel expectedResult = mock(ResultModel.class);
-    doReturn(expectedResult).when(commandDelegate).executeStopGatewaySender(any(), any(), any());
-
-    StopGatewaySenderCommand command =
-        new StopGatewaySenderCommand(commandDelegateSupplier, findMembers);
+    String gatewaySenderIsStoppedMsg = "GatewaySender ln is stopped on member";
+    List<Future<List<String>>> futures = new ArrayList<>();
+    for (int memberIndex = 0; memberIndex < members.size(); memberIndex++) {
+      Future<List<String>> future = mock(Future.class);
+      List<String> list = Arrays.asList("member" + memberIndex, "OK", gatewaySenderIsStoppedMsg);
+      doReturn(list).when(future).get();
+      futures.add(future);
+    }
+    doReturn(futures).when(executorService).invokeAll(any());
 
     // act
-    ResultModel commandResult =
-        gfsh.executeAndAssertThat(command, "stop gateway-sender --id=sender1").getCommandResult()
-            .getResultData();
+    StopGatewaySenderCommand command = new StopGatewaySenderCommand(executorService,
+        stopperOnMemberFactory, managementService, findMembers);
+    ResultModel result = command.executeStopGatewaySender(senderId, cache, members);
 
     // assert
-    assertThat(commandResult).isEqualTo(expectedResult);
+    assertThat(result.isSuccessful()).isTrue();
+
+    TabularResultModel resultData = result.getTableSection(CliStrings.STOP_GATEWAYSENDER);
+    List<String> member = resultData.getValuesInColumn("Member");
+    assertThat(member).containsExactlyInAnyOrder("member0", "member1", "member2");
+    List<String> status = resultData.getValuesInColumn("Result");
+    assertThat(status).containsExactlyInAnyOrder("OK", "OK", "OK");
+    List<String> message = resultData.getValuesInColumn("Message");
+    assertThat(message).containsExactlyInAnyOrder(gatewaySenderIsStoppedMsg,
+        gatewaySenderIsStoppedMsg, gatewaySenderIsStoppedMsg);
+
+    ArgumentCaptor<Collection> callablesCaptor =
+        ArgumentCaptor.forClass(Collection.class);
+    verify(executorService, times(1)).invokeAll((callablesCaptor.capture()));
+    assertThat(callablesCaptor.getValue().size()).isEqualTo(members.size());
+    verify(executorService, times(1)).shutdown();
   }
+
+  @SuppressWarnings("unchecked")
+  @Test
+  public void stopGatewaySenderInterruptedReturnsError() throws InterruptedException {
+    // arrange
+    InterruptedException exception = new InterruptedException("interruption2");
+    doThrow(exception).when(executorService).invokeAll(any());
+
+    // act
+    StopGatewaySenderCommand command =
+        new StopGatewaySenderCommand(executorService, stopperOnMemberFactory,
+            managementService, findMembers);
+    ResultModel result = command.executeStopGatewaySender(senderId, cache, members);
+
+    // assert
+    assertThat(result.isSuccessful()).isFalse();
+    assertThat(result.getInfoSection("info").getContent().get(0)).isEqualTo(
+        "Could not invoke stop gateway sender sender1 operation on members due to interruption2");
+  }
+
 }
